backend/requirements.txt
firebase-admin==5.3.0
Flask==2.2.2
google-cloud-aiplatform==1.16.1
gunicorn==20.1.0
Pillow==9.2.0
requests==2.28.1
